refactor: extract index-row support into a mixin
The five methods under the "Support for PdsFile objects representing index rows"
banner move verbatim into src/pdsfile/_index_rows.py as _IndexRowsMixin, which
the class statement takes as a base in its alphabetical slot. The bodies are
byte-for-byte what they were; nothing else in this commit is a content edit.

The free-variable sweep names three module-level references: numbers, pdstable
and _clean_join. numbers and pdstable have no other user in pdsfile.py once the
block leaves, but both are frozen members of pdsfile.pdsfile's public surface, so
their imports stay in the redundant-alias re-export form the header already uses
for six modules.

The new module's per-file-ignores entry is a split of pdsfile.py's: RUF005 and
UP024 are the codes its moved lines carry, both already forgiven for the same
lines in the same file, and every code in pdsfile.py's entry conserves.
